+++ engine/flink/management/periodic/src/main/scala/pl/touk/nussknacker/engine/management/periodic/flink/FlinkJarManager.scala
@@ -5,9 +5,9 @@ import pl.touk.nussknacker.engine.ModelData
 import pl.touk.nussknacker.engine.api.ProcessVersion
 import pl.touk.nussknacker.engine.api.deployment.{DeploymentData, ExternalDeploymentId, GraphProcess}
 import pl.touk.nussknacker.engine.management.periodic.model.DeploymentWithJarData
-import pl.touk.nussknacker.engine.management.periodic.{EnrichDeploymentWithJarData, JarManager, PeriodicBatchConfig, model}
+import pl.touk.nussknacker.engine.management.periodic.{DeploymentEnricher, DeploymentEnricherInputData, JarManager, PeriodicBatchConfig, model}
 import pl.touk.nussknacker.engine.management.rest.{FlinkClient, HttpFlinkClient}
-import pl.touk.nussknacker.engine.management.{FlinkConfig, FlinkModelJar, FlinkDeploymentManager, FlinkStreamingRestManager}
+import pl.touk.nussknacker.engine.management.{FlinkConfig, FlinkDeploymentManager, FlinkModelJar, FlinkStreamingRestManager}
 import pl.touk.nussknacker.engine.modelconfig.InputConfigDuringExecution
 import sttp.client.{NothingT, SttpBackend}
 
@@ -19,14 +19,14 @@ private[periodic] object FlinkJarManager {
   def apply(flinkConfig: FlinkConfig,
             periodicBatchConfig: PeriodicBatchConfig,
             modelData: ModelData,
-            enrichDeploymentWithJarData: EnrichDeploymentWithJarData)
+            deploymentEnricher: DeploymentEnricher)
            (implicit backend: SttpBackend[Future, Nothing, NothingT], ec: ExecutionContext): JarManager = {
     new FlinkJarManager(
       flinkClient = new HttpFlinkClient(flinkConfig),
       jarsDir = Paths.get(periodicBatchConfig.jarsDir),
       modelConfig = modelData.inputConfigDuringExecution,
       createCurrentModelJarFile = new FlinkModelJar().buildJobJar(modelData),
-      enrichDeploymentWithJarData = enrichDeploymentWithJarData
+      deploymentEnricher = deploymentEnricher
     )
   }
 }
@@ -36,7 +36,7 @@ private[periodic] class FlinkJarManager(flinkClient: FlinkClient,
                                         jarsDir: Path,
                                         modelConfig: InputConfigDuringExecution,
                                         createCurrentModelJarFile: => File,
-                                        enrichDeploymentWithJarData: EnrichDeploymentWithJarData)
+                                        deploymentEnricher: DeploymentEnricher)
   extends JarManager with LazyLogging {
 
   import scala.concurrent.ExecutionContext.Implicits.global
@@ -46,15 +46,15 @@ private[periodic] class FlinkJarManager(flinkClient: FlinkClient,
   override def prepareDeploymentWithJar(processVersion: ProcessVersion,
                                         processJson: String): Future[DeploymentWithJarData] = {
     logger.info(s"Prepare deployment for scenario: $processVersion")
-    copyJarToLocalDir(processVersion).flatMap { jarFileName =>
-      val deploymentWithJarData = model.DeploymentWithJarData(
-        processVersion = processVersion,
-        processJson = processJson,
-        modelConfig = modelConfig.serialized,
-        jarFileName = jarFileName
-      )
-      enrichDeploymentWithJarData(deploymentWithJarData)
-    }
+    for {
+      jarFileName <- copyJarToLocalDir(processVersion)
+      enrichedDeployment <- deploymentEnricher.onInitialSchedule(DeploymentEnricherInputData(processJson = processJson, modelConfig = modelConfig.serialized))
+    } yield DeploymentWithJarData(
+      processVersion = processVersion,
+      processJson = processJson,
+      modelConfig = enrichedDeployment.modelConfig,
+      jarFileName = jarFileName
+    )
   }
 
   private def copyJarToLocalDir(processVersion: ProcessVersion): Future[String] = Future {
@@ -70,11 +70,15 @@ private[periodic] class FlinkJarManager(flinkClient: FlinkClient,
     val processVersion = deploymentWithJarData.processVersion
     logger.info(s"Deploying scenario ${processVersion.processName.value}, version id: ${processVersion.versionId} and jar: ${deploymentWithJarData.jarFileName}")
     val jarFile = jarsDir.resolve(deploymentWithJarData.jarFileName).toFile
-    val args = FlinkDeploymentManager.prepareProgramArgs(deploymentWithJarData.modelConfig,
-      processVersion,
-      deploymentData,
-      GraphProcess(deploymentWithJarData.processJson))
-    flinkClient.runProgram(jarFile, FlinkStreamingRestManager.MainClassName, args, None)
+    val deploymentEnricherInputData = DeploymentEnricherInputData(processJson = deploymentWithJarData.processJson, modelConfig = deploymentWithJarData.modelConfig)
+    for {
+      enrichedDeployment <- deploymentEnricher.onDeploy(deploymentEnricherInputData)
+      args = FlinkDeploymentManager.prepareProgramArgs(enrichedDeployment.modelConfig,
+        processVersion,
+        deploymentData,
+        GraphProcess(deploymentWithJarData.processJson))
+      externalDeploymentId <- flinkClient.runProgram(jarFile, FlinkStreamingRestManager.MainClassName, args, None)
+    } yield externalDeploymentId
   }
 
   override def deleteJar(jarFileName: String): Future[Unit] = {